move blockarrays to extension
