+++ Project.toml
@@ -4,8 +4,13 @@ authors = ["ITensor developers <support@itensor.org> and contributors"]
 version = "0.1.0"
 
 [deps]
-BlockArrays = "8e7c35d0-a365-5155-bbbb-fb81a777f24e"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+
+[weakdeps]
+BlockArrays = "8e7c35d0-a365-5155-bbbb-fb81a777f24e"
+
+[extensions]
+LabelledNumbersBlockArraysExt = "BlockArrays"
 
 [compat]
 BlockArrays = "1.2.0"
