+++ src/LabelledNumbers.jl
@@ -1,8 +1,9 @@
 module LabelledNumbers
+
 include("labelled_interface.jl")
 include("labellednumber.jl")
 include("labelledinteger.jl")
 include("labelledarray.jl")
 include("labelledunitrange.jl")
-include("LabelledNumbersBlockArraysExt.jl")
+
 end
